Clean up go modules fix
- didn't need to change GO111MODULE var
- update checksum for go1.11.2

+++ scripts/install_go.sh
@@ -4,23 +4,21 @@ set -euo pipefail
 GO_VERSION="1.11.2"
 
 export GoInstallDir="/tmp/go$GO_VERSION"
-# export GO1111MODULE=on
 mkdir -p $GoInstallDir
 
 if [ ! -f $GoInstallDir/go/bin/go ]; then
-  GO_MD5="0571886e9b9ba07773b542a11e9859a4"
-  # URL=https://buildpacks.cloudfoundry.org/dependencies/go/go${GO_VERSION}.linux-amd64-${GO_MD5:0:8}.tar.gz
-  URL=https://dl.google.com/go/go1.11.2.linux-amd64.tar.gz
+  GO_SHA256="1dfe664fa3d8ad714bbd15a36627992effd150ddabd7523931f077b3926d736d"
+  URL=https://dl.google.com/go/go${GO_VERSION}.linux-amd64.tar.gz
 
   echo "-----> Download go ${GO_VERSION}"
   curl -s -L --retry 15 --retry-delay 2 $URL -o /tmp/go.tar.gz
 
-  # DOWNLOAD_MD5=$(md5sum /tmp/go.tar.gz | cut -d ' ' -f 1)
+  DOWNLOAD_SHA256=$(shasum -a 256 /tmp/go.tar.gz | cut -d ' ' -f 1)
 
-  # if [[ $DOWNLOAD_MD5 != $GO_MD5 ]]; then
-  #   echo "       **ERROR** MD5 mismatch: got $DOWNLOAD_MD5 expected $GO_MD5"
-  #   exit 1
-  # fi
+  if [[ $DOWNLOAD_SHA256 != $GO_SHA256 ]]; then
+    echo "       **ERROR** SHA256 mismatch: got $DOWNLOAD_SHA256 expected $GO_SHA256"
+    exit 1
+  fi
 
   tar xzf /tmp/go.tar.gz -C $GoInstallDir
   rm /tmp/go.tar.gz
